requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
faiss-cpu==1.7.4
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
requests==2.31.0  # <--- Thêm dòng này
gdown==4.7.1      # <--- Thêm dòng này